+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/BoldTextMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.BoldTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.boldHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.boldMarkdown
 
@@ -7,7 +9,4 @@ data class BoldTextMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.boldMarkdown()
-    override val asHtmlSource: String = sourceString.boldHTML()
-}
+) : MessageEntity, TextSource by BoldTextSource(sourceString)