+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/BotCommandMessageEntity.kt
@@ -1,19 +1,16 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.BotCommandTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.commandHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.commandMarkdown
-
-private val commandRegex = Regex("[/!][^@\\s]*")
 
 data class BotCommandMessageEntity(
     override val offset: Int,
     override val length: Int,
-    override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.commandMarkdown()
-    override val asHtmlSource: String = sourceString.commandHTML()
-
-    val command: String by lazy {
-        commandRegex.find(sourceString) ?.value ?.substring(1) ?: sourceString.substring(1)// skip first symbol like "/" or "!"
-    }
+    override val sourceString: String,
+    private val botCommandTextSource: BotCommandTextSource = BotCommandTextSource(sourceString)
+) : MessageEntity, TextSource by botCommandTextSource {
+    val command: String
+        get() = botCommandTextSource.command
 }