+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/HashTagMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.HashTagTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.hashTagHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.hashTagMarkdown
 
@@ -7,7 +9,4 @@ data class HashTagMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.hashTagMarkdown()
-    override val asHtmlSource: String = sourceString.hashTagHTML()
-}
+) : MessageEntity, TextSource by HashTagTextSource(sourceString)