+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/PreTextMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.PreTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.preHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.preMarkdown
 
@@ -7,7 +9,4 @@ data class PreTextMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.preMarkdown()
-    override val asHtmlSource: String = sourceString.preHTML()
-}
+) : MessageEntity, TextSource by PreTextSource(sourceString)