+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/RegularTextMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.RegularTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.toHtml
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.toMarkdown
 
@@ -7,7 +9,4 @@ data class RegularTextMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.toMarkdown()
-    override val asHtmlSource: String = sourceString.toHtml()
-}
+) : MessageEntity, TextSource by RegularTextSource(sourceString)