+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/URLMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.URLTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.linkHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.linkMarkdown
 
@@ -7,9 +9,6 @@ data class URLMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity{
+) : MessageEntity, TextSource by URLTextSource(sourceString) {
     val url: String = sourceString
-
-    override val asMarkdownSource: String = sourceString.linkMarkdown(url)
-    override val asHtmlSource: String = sourceString.linkHTML(url)
 }