+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/textsources/EMailTextSource.kt
@@ -1,0 +1,12 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources
+
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.utils.emailHTML
+import com.github.insanusmokrassar.TelegramBotAPI.utils.emailMarkdown
+
+class EMailTextSource(
+    sourceString: String
+) : TextSource {
+    override val asMarkdownSource: String = sourceString.emailMarkdown()
+    override val asHtmlSource: String = sourceString.emailHTML()
+}